Add instructional text to intro narration, etc.
Replaced CONTROLS_NARRATION with INSTRUCTIONS_NARRATION_1 and added
INSTRUCTIONS_NARRATION_2, both of which are displayed when the game is
played for the first time as well as when "Instructions" (formerly
"Controls") is selected from the main menu. Also improved narration
transition functionality by causing all buttons to bring up the next
narration.